# go/APIvidgm/docs/swagger.yaml
definitions:
...
  models.Review:
    properties:
      comment:
        type: string
      id:
        type: integer
      modstatus:
        description: pending, approved, rejected
        type: string
      rating:
        type: integer
      user_id:
        type: integer
      username:
        type: string
      videogame_id:
        type: integer
      videogame_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Shit documentation for shit API.
  termsOfService: http://swagger.io/terms/
  title: VideoGames API
  version: "1.0"
paths:
...
  /reviews:
    get:
      consumes:
      - application/json
      description: Retrieves all reviews from all users for all videogames. Requires
        appropriate access if restricted.
      produces:
      - application/json
      responses:
        "200":
          description: List of all reviews
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
        "404":
          description: Error retrieving reviews
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retrieve all reviews
      tags:
      - Reviews
  /reviews/{id}/approve:
    patch:
      consumes:
      - application/json
      description: Allows an admin to approve a pending review, changing its moderation
        status to "approved".
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Review approved successfully
          schema:
            type: string
        "403":
          description: Forbidden – insufficient privileges
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Approve a user review
      tags:
      - Reviews
  /reviews/{id}/reject:
    patch:
      consumes:
      - application/json
      description: Allows an admin to reject a pending review, changing its moderation
        status to "rejected".
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Review rejected successfully
          schema:
            type: string
        "403":
          description: Forbidden – insufficient privileges
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Reject a user review
      tags:
      - Reviews
...
  /users/{id}/reviews:
    get:
      consumes:
      - application/json
      description: Retrieves all reviews created by a specific user. Requires authentication
        as that user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of user reviews
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all reviews made by a user
      tags:
      - Reviews
  /users/{id}/reviews/{gameID}:
    get:
      consumes:
      - application/json
      description: Retrieves the review written by a specific user for a given videogame.
        Requires authentication as that user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Videogame ID
        in: path
        name: gameID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User review for the videogame
          schema:
            $ref: '#/definitions/models.Review'
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User, videogame, or review not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user's review for a specific videogame
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Allows a user to create a new review or update an existing one
        for a specific videogame. Requires authentication as that user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Videogame ID
        in: path
        name: gameID
        required: true
        type: integer
      - description: Review content (rating, comment, etc.)
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "200":
          description: Created or updated review
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Invalid input or missing parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User or videogame not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Database or internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create or update a user's review for a videogame
      tags:
      - Reviews
  /users/{id}/reviews/{reviewID}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific review made by a user. The authenticated user
        must be the owner of the review.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Review successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Review does not belong to user or unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User or review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user review
      tags:
      - Reviews
...
  /videogames/{id}/reviews:
    get:
      consumes:
      - application/json
      description: Retrieves all reviews associated with a specific videogame by its
        ID.
      parameters:
      - description: Videogame ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of reviews for the videogame
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
        "404":
          description: Videogame not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all reviews for a videogame
      tags:
      - Reviews
swagger: "2.0"